+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #575 Better contextual message for Cayenne ValidationException
 * #577 Better default encoder for DataResponse
+* #578 Overlaying AgIdPart access rules results in their renaming
 
 ## Release 5.0.M6