+++ RELEASE-NOTES.md
@@ -1,3 +1,7 @@
+## Release 5.0.M8
+* 
+* #583 Upgrade jackson-databind to 2.13.4.2
+
 ## Release 5.0.M7
 
 * #575 Better contextual message for Cayenne ValidationException